fox_pos_project/requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==23.5
django-cors-headers==4.3.1
Pillow==10.1.0
python-barcode==0.15.1
reportlab==4.0.7
openpyxl==3.1.2
pandas==2.1.4
python-dateutil==2.8.2
